[STABLE] Factor assets and store concerns into their own files
