+++ src/store.js
@@ -2,6 +2,8 @@ import { compose, createStore } from 'redux';
 import middleware from './middleware';
 import reducers from './reducers';
 
-export default function configureStore() {
+function configureStore() {
   return compose(middleware)(createStore)(reducers);
 }
+
+export var store = configureStore();
